refactor(notes): service files
