+++ src/modules/notes/Note/components/HeaderNote/services/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./fetchPostCreateNoteService.types";
