+++ src/modules/notes/Note/components/HeaderNote/types/HeaderNote.types.ts
@@ -1,6 +1,5 @@
 import { IFetchGetOneBoardResponse } from "shared/common/hook/useFetchGetOneBoard/types";
 import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
 
 export interface IHeaderNoteView {
 	data: IFetchGetOneBoardResponse | undefined;
@@ -10,19 +9,4 @@ export interface IHeaderNoteView {
 	countNotes: number;
 	isLoading: boolean;
 	isFetching: boolean;
-}
-
-export interface IFetchPostCreateNoteResponse extends IFetchResponseDefault {
-	results: {
-		note_id: string;
-	};
-}
-
-export interface IFetchPostCreateNoteService {
-	board_id: string | undefined;
-	payload: {
-		color_note: string;
-		title_note: string;
-		observation: string;
-	};
 }
